requirements.txt
#
##### SToPA requirements.txt
#

deskew
fuzzywuzzy
gitpython
jupyterlab
matplotlib
notebook
pandas
pyarrow
PyPDF2
python-Levenshtein
usaddress 


